impl get data, fix schema sql

Closes #6

See merge request teepij/whereismyhome_vueback!9

Former-commit-id: dda1f2ddb6ec4a6e4e937c35827fc077d83a8fe4

+++ src/main/resources/sql/securityindex_schema.sql
@@ -5,7 +5,7 @@ USE `vueproject` ;
 DROP TABLE IF EXISTS `vueproject`.`securityindex` ;
 
 CREATE TABLE IF NOT EXISTS `vueproject`.`securityindex` (
-  `sigugunCode` INT NOT NULL,
+  `sigugunCode` BIGINT NOT NULL,
 `sido` VARCHAR(30) NOT NULL,
 `city` VARCHAR(30) NOT NULL,
 `region_name` VARCHAR(30) NOT NULL,
